src/preprocess/segment.py
import torch
import numpy as np
from PIL import Image
import cv2

class Segmenter:
    def __init__(self):
        # This would ideally load MODNet or a similar model
        print("Segmenter initialized. Ready for background removal.")

    def remove_background(self, pil_image):
        # Placeholder: returning image as is
        # Actual implementation would use a mask generation model
        print("Removing background...")
        return pil_image

    def get_face_mask(self, pil_image):
        # Generate a segmentation mask for facial parts
        print("Generating face segmentation mask...")
        # For now, return a dummy mask
        w, h = pil_image.size
        return Image.fromarray(np.ones((h, w), dtype=np.uint8) * 255)

if __name__ == "__main__":
    segmenter = Segmenter()
    print("Segmenter module ready.")
