requirements.txt
torch
torchvision
torchaudio
mediapipe==0.10.14
opencv-python
numpy
pillow
scipy
gdown
face-alignment
chumpy
pyrender
trimesh
flask
fastapi
uvicorn
python-multipart
gradio
diffusers
transformers
accelerate
controlnet-aux

onnxruntime-gpu
tqdm
matplotlib
pytest
